+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- Add bundled ChimeraX script recipe search/read tools with rich-report payload examples.
+
 ## [0.3.0] - 2026-05-20
 
 ### Added